cyan/requirements.txt
docker-py
python-etcd
